+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 # Changelog
 
 All notable changes to `vuedata` will be documented in this file.
+
+## V1.0.2
+When writing the data-block, comments in the data-block are ignored
+
+## V1.0.1
+When reading the data-block, comments in the data-block are ignored